+++ lib/app_review.dart
@@ -1,13 +1,14 @@
 import 'dart:async';
-import 'package:http/http.dart' as http;
-import 'package:flutter/services.dart';
-import 'dart:io';
 import 'dart:convert';
-import 'package:url_launcher/url_launcher.dart';
+import 'dart:io';
+
+import 'package:flutter/services.dart';
+import 'package:http/http.dart' as http;
 import 'package:package_info/package_info.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class AppReview {
-  static const MethodChannel _channel = const MethodChannel('app_review');
+  static const MethodChannel _channel = MethodChannel('app_review');
 
   static Future<String> get requestReview async {
     if (Platform.isIOS) {
@@ -21,9 +22,9 @@ class AppReview {
 
   static Future<String> get writeReview async {
     if (Platform.isIOS) {
-      String _appID = await getiOSAppID;
+      final String _appID = await getiOSAppID;
       String details = '';
-      String _url =
+      final String _url =
           'itunes.apple.com/us/app/id$_appID?mt=8&action=write-review';
       if (await canLaunch("itms-apps://")) {
         print('launching store page');
@@ -43,11 +44,11 @@ class AppReview {
   static Future<String> get storeListing async {
     String details = '';
     if (Platform.isIOS) {
-      String _appID = await getiOSAppID;
+      final String _appID = await getiOSAppID;
       await launch('https://itunes.apple.com/us/app/id$_appID?');
       details = 'Launched App Store';
     } else {
-      String _appID = await getAppID;
+      final String _appID = await getAppID;
       if (await canLaunch("market://")) {
         print('launching store page');
         await launch("market://details?id=" + _appID);
@@ -61,12 +62,12 @@ class AppReview {
   }
 
   static Future<String> get getAppID async {
-    PackageInfo packageInfo = await PackageInfo.fromPlatform();
+    final PackageInfo packageInfo = await PackageInfo.fromPlatform();
 
-    String appName = packageInfo.appName;
-    String packageName = packageInfo.packageName;
-    String version = packageInfo.version;
-    String buildNumber = packageInfo.buildNumber;
+    final String appName = packageInfo.appName;
+    final String packageName = packageInfo.packageName;
+    final String version = packageInfo.version;
+    final String buildNumber = packageInfo.buildNumber;
 
     print(
         "App Name: $appName\nPackage Name: $packageName\nVersion: $version\nBuild Number: $buildNumber");
@@ -79,8 +80,8 @@ class AppReview {
     String _id = '';
     await http
         .get('http://itunes.apple.com/lookup?bundleId=$_appID')
-        .then((response) {
-      Map<String, dynamic> _json = json.decode(response.body);
+        .then((dynamic response) {
+      final Map<String, dynamic> _json = json.decode(response.body);
       _id = _json['results'][0]['trackId'].toString();
       print('Track ID: $_id');
     });